Find the approximate set of neighbors of each voxels.

Find the approximate set of neighborhood of each voxels, given the distance radius.The number of elements in each neighborhood is (2*radius)^3. We recommend to find the nearest neighbors prior to limma

Usage

findNearestNeighbors(maskImg, radius = 1, threads = parallel::detectCores())

Arguments

maskImg

Input Masked image which the masked voxels are labeled with zeros

radius

Integer: Indices that `dist(targetInd - queryInd) <= radius`

threads

Integer: threads to use in closest point search.

Value

The result from vcgKDtree