Run `eBayes` on NIfTI Images

Run `eBayes` on NIfTI Images

Usage

nifti_eBayes(
  imgs,
  mask = NULL,
  verbose = TRUE,
  ...,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  coef = NULL,
  adjust.method = "BH"
)

eBayes_to_images(
  fit,
  mask,
  verbose = TRUE,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  coef = NULL,
  adjust.method = "BH"
)

Arguments

imgs

List of images or character of filenames

mask

Image to subset the data

verbose

Print diagnostic messages

...

additional arguments to pass to lmFit

proportion

numeric value between 0 and 1, assumed proportion of voxels which are differentially expressed, passed to eBayes

stdev.coef.lim

numeric vector of length 2, assumed lower and upper limits for the standard deviation of log2-fold-changes for differentially expressed voxels, passed to eBayes

trend

logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance is constant, passed to eBayes

robust

logical, should the estimation of df.prior and var.prior be robustified against outlier sample variances, passed to eBayes

winsor.tail.p

numeric vector of length 1 or 2, giving left and right tail proportions of x to Winsorize. Used only when robust=TRUE, passed to eBayes

coef

column number or column name specifying which coefficient or contrast of the linear model is of interest, passed to topTable

adjust.method

method to adjust p-value, passed to topTable

fit

an MArrayLM fitted model object produced by `lmFit` or `contrasts.fit`.

Value

A list of output images from `eBayes` and the linear fit

Examples

tarfile = system.file("extdata", "can.tar.gz", package = "limmi")
tarfile
#> [1] "/home/runner/work/_temp/Library/limmi/extdata/can.tar.gz"
exdir = tempfile()
dir.create(exdir)
files = untar(tarfile = tarfile, list = TRUE, exdir = exdir)
files = files[!grepl("^[.]", basename(files))]
unz = untar(tarfile = tarfile, files = files, exdir = exdir)
files = files[ grepl("hdr", basename(files))]
files
#>  [1] "con_0006.hdr" "con_0007.hdr" "con_0008.hdr" "con_0009.hdr" "con_0010.hdr"
#>  [6] "con_0011.hdr" "con_0012.hdr" "con_0013.hdr" "con_0014.hdr" "con_0015.hdr"
#> [11] "con_0016.hdr" "con_0017.hdr"
files = file.path(exdir, files)
res = nifti_eBayes(files, verbose = FALSE)
if (FALSE) { # \dontrun{
design_mat = cbind(x1 = rep(c(0, 1), each = 6),
x2 = rep(c(0, 1, 0, 1), each = 3))
res2 = nifti_eBayes(files, design = design_mat, coef = 2)
} # }