Transform set of images to matrix

Creates a matrix, where the voxels are on the rows and images are on the columns

Usage

nifti_images_to_matrix(imgs, mask = NULL, verbose = TRUE)

Arguments

imgs

Vector of files or list of images (niftiImage, array, or nifti)

mask

Binary image to subset the voxels

verbose

print diagnostic messages

Value

Matrix of V by n, where V is the product of the dimensions of one image or the number of voxels in the mask, and n is the number of images