Convert NIfTI data.frame to a SummarizedExperiment

Convert NIfTI data.frame to a SummarizedExperiment

Usage

nifti_df_to_SummarizedExperiment(
  x,
  path_column = "image",
  assay_name = path_column,
  template_images = NULL,
  ...
)

Arguments

x

An optional `data.frame` describing the samples/participants.

path_column

column in `x` that is the paths to the file names of the data. Can also be a list column of `x` where this column are the images

assay_name

Name of the `assay` in the SummarizedExperiment

template_images

List of images that are annotations of the rows of the data (voxels), such as template mask or segmentations

...

additional arguments to pass to SummarizedExperiment

Value

A SummarizedExperiment output

Examples

tarfile = system.file("extdata", "can.tar.gz", package = "limmi")
tarfile
#> [1] "/home/runner/work/_temp/Library/limmi/extdata/can.tar.gz"
exdir = tempfile()
dir.create(exdir)
files = untar(tarfile = tarfile, list = TRUE, exdir = exdir)
files = files[!grepl("^[.]", basename(files))]
unz = untar(tarfile = tarfile, files = files, exdir = exdir)
files = files[ grepl("hdr", basename(files))]
files
#>  [1] "con_0006.hdr" "con_0007.hdr" "con_0008.hdr" "con_0009.hdr" "con_0010.hdr"
#>  [6] "con_0011.hdr" "con_0012.hdr" "con_0013.hdr" "con_0014.hdr" "con_0015.hdr"
#> [11] "con_0016.hdr" "con_0017.hdr"
files = file.path(exdir, files)
df = data.frame(image = files,
age = stats::rpois(length(files), 50))
se = nifti_df_to_SummarizedExperiment(df, "image")
#>   |                                                                              |                                                                      |   0%  |                                                                              |======                                                                |   8%  |                                                                              |============                                                          |  17%  |                                                                              |==================                                                    |  25%  |                                                                              |=======================                                               |  33%  |                                                                              |=============================                                         |  42%  |                                                                              |===================================                                   |  50%  |                                                                              |=========================================                             |  58%  |                                                                              |===============================================                       |  67%  |                                                                              |====================================================                  |  75%  |                                                                              |==========================================================            |  83%  |                                                                              |================================================================      |  92%  |                                                                              |======================================================================| 100%